feat: add option to hydrate when entering the viewport
