+++ src/runtime/components/DelayHydration.ts
@@ -2,20 +2,20 @@
 // @see https://github.com/harlan-zw/nuxt-delay-hydration/blob/main/src/template/global.ts
 // @license MIT
 
-import { createStaticVNode, ref, onMounted, defineComponent, getCurrentInstance } from 'vue'
+import { createStaticVNode, ref, defineComponent, getCurrentInstance } from 'vue'
 import type { VNode } from 'vue'
+import { isIntersectionObserverSupported, useViewportListener, useWindowListeners } from '../composables'
 import { useNuxtApp } from '#imports'
 import { getFragmentHTML } from '#app/components/utils'
-import { delayHydrationOptions } from '#build/module/nuxt-vitalizer'
-
-interface Handler {
-  // eslint-disable-next-line @typescript-eslint/no-invalid-void-type
-  promise: Promise<void | Event>
-  cleanup: () => void
-}
 
 export default defineComponent({
-  setup(_, { slots }) {
+  props: {
+    hydrateWhenVisible: {
+      type: Boolean,
+      default: false,
+    },
+  },
+  setup(props, { slots }) {
     const nuxtApp = useNuxtApp()
     const instance = getCurrentInstance()!
     const supportsIdleCallback = import.meta.client ? ('requestIdleCallback' in window) : false
@@ -28,63 +28,14 @@ export default defineComponent({
         vnode = createStaticVNode(fragment.join(''), fragment.length)
       }
 
-      onMounted(async () => {
-        const triggers = [idleListener(), eventListeners()]
-        nuxtApp._delayHydrationPromise ??= Promise.race(
-          triggers.map(t => t.promise),
-        ).finally(() => {
-          for (const t of triggers) t.cleanup()
-        })
-
-        await nuxtApp._delayHydrationPromise
-        shouldRender.value = true
-      })
+      if (instance.vnode.el && props.hydrateWhenVisible && isIntersectionObserverSupported()) {
+        useViewportListener(instance.vnode.el, shouldRender)
+      }
+      else {
+        useWindowListeners(shouldRender)
+      }
     }
 
     return () => (shouldRender.value ? slots.default?.() : vnode)
   },
 })
-
-function eventListeners(): Handler {
-  const abortController = new AbortController()
-  const promise = new Promise<Event>((resolve) => {
-    const listener = (e: Event) => {
-      for (const e of delayHydrationOptions.hydrateOnEvents) window.removeEventListener(e, listener)
-      requestAnimationFrame(() => resolve(e))
-    }
-
-    for (const e of delayHydrationOptions.hydrateOnEvents) {
-      window.addEventListener(e, listener, {
-        capture: true,
-        once: true,
-        passive: true,
-        signal: abortController.signal,
-      })
-    }
-  })
-
-  return {
-    promise,
-    cleanup: () => abortController.abort(),
-  }
-}
-
-function idleListener(): Handler {
-  let idleId: number
-
-  const promise = new Promise<void>((resolve) => {
-    const timeoutDelay = () => {
-      setTimeout(() => {
-        requestAnimationFrame(() => resolve())
-      }, delayHydrationOptions.postIdleTimeout)
-    }
-    idleId = requestIdleCallback(timeoutDelay, {
-      timeout: delayHydrationOptions.idleCallbackTimeout,
-    })
-  })
-
-  return {
-    promise,
-    cleanup: () => cancelIdleCallback(idleId),
-  }
-}
