+++ src/runtime/composables/index.ts
@@ -1,0 +1,6 @@
+export * from './use-viewport-listener'
+export * from './use-window-listeners'
+
+export function isIntersectionObserverSupported() {
+  return window && 'IntersectionObserver' in window
+}
